Pull out local parametric studies.
