+++ devel/undocumented_fields/undocumented_fields.txt
@@ -120,7 +120,7 @@
   Class: StandaloneLauncher
     - file_transfer_service
 
-  Module: ansys.fluent.core.legacy.local_parametric_study
+  Module: ansys.fluent.core.local_parametric_study
   Class: LocalDesignPoint
     - name
   Class: LocalParametricStudy
