+++ tests/parametric/test_local_parametric_run.py
@@ -24,7 +24,7 @@
 import pytest
 
 from ansys.fluent.core import examples
-from ansys.fluent.core.legacy.local_parametric_study import (
+from ansys.fluent.core.local_parametric_study import (
     LocalParametricStudy,
     convert_design_point_parameter_units,
 )
